+++ models/Delay.cs
@@ -2,11 +2,11 @@ using System;
 
 namespace LondonTube {
   class Delay {
-    public int Time {get;}
+    public Double Time {get;}
     public TrackDelay TrackDelay {get;}
     public Connection Connection {get;}
 
-    public Delay(Connection connection, TrackDelay trackDelay, int time){
+    public Delay(Connection connection, TrackDelay trackDelay, Double time){
       
       Connection = connection;
       TrackDelay = trackDelay;
